+++ types/github.ts
@@ -1,0 +1,17 @@
+import { RestEndpointMethodTypes } from "@octokit/plugin-rest-endpoint-methods";
+
+export type GitHubIssue = RestEndpointMethodTypes["issues"]["get"]["response"]["data"];
+export type GitHubLabel = RestEndpointMethodTypes["issues"]["listLabelsOnIssue"]["response"]["data"][0];
+
+export type StateChanges<T extends string = "open" | "closed"> = {
+    [key: string]: {
+        cause: boolean;
+        effect: T;
+        comment: string;
+    };
+};
+
+export enum LABELS {
+    PRICE = "Price",
+    UNAVAILABLE = "Unavailable",
+}
